+++ CHANGELOG.md
@@ -7,7 +7,6 @@
 
 * Added github URL's to package.json
 
-
 ### 1.0.2 (2021-05-05)
 
 * Documentation Added
@@ -15,3 +14,8 @@
 ### 1.0.3 (2022-07-04)
 
 * Updated error handling
+
+### 2.0.0 (2022-10-25)
+
+* SDK v2.0.0 - This SDK can now detect and return the details of all the NFTs for a particular address across any supported chains.
+* Supported Chains : `Ethereum`, `Polygon` and `BSC`.